refactor(scripts): extract the index resolver and the pseudonym logic
AGENTS.md rule 3: when you edit inside a file at the size limit, extract the
region you are working on and leave the I/O behind. Milestone 3 pushed
`fastq_bam_processing.py` to 669 lines, past the ~650 guideline, and grew
`cohort_inputs.py` by 219.

`alignment_index.py` takes `resolve_bam_index`, the htslib resolution order
with no samtools call in it. `cohort_pseudonyms.py` takes the digest
constants and `pseudonymized_sample_name` with its validation; it imports
nothing but hashlib, logging and typing.

`duplicate_identity` deliberately stays in `cohort_inputs.py`. It runs
upstream of any digest -- two directories both named `sample` are two
patients with one basename, and no digest width separates them -- and it
reads discovery's own invariant that two entries always come from two
distinct directories. Filing it under pseudonyms would say the opposite of
what it does.

Both new modules are fully annotated and at 100% branch coverage, so the
Layout section's claim about the extracted modules still holds with them in
it. `fastq_bam_processing.py` is back to 648.
